Got itoa() working???
Added the prolog and epilog to the itoa function and cleaned up the input parameters. Also, re-wrote part of main.s to test that itoa function is workingl. At this time, given the very simple test, it does appear to be working. Next, I need to go back and clean the source of itoa.s and do some basic validation that the itoa function does indeed work.