Refactored itoa function
Refactored itoa function so that it makes more streamlined use of the new pow function and doesn't have a lot of ugly boiler plate and bloat wrapped around the pow function call. In redoing the itoa function I also refrained from using regesters 0-3 for persistant storage to be more in line with the linux arm32 abi. That said, the file needs to be cleaned up again and tested more before moving on.